refactor: Outbox 관련 데이터 추가 및 개선 #4
